producer/requirements.txt
kafka-python-ng
fastapi
uvicorn
pydantic